Move MarketActivityRecord to new namespace
